# shared/contracts/openapi/fudcourt.yaml
openapi: 3.0.3
info:
  title: FUDCourt Executor API
  version: 0.1.0
  description: |
    The **executor HTTP surface** of FUDCourt as it exists TODAY.

    This contract documents the CURRENT TS-owned surface during the Go
    migration: the ground truth is the frozen contract in
    `frontend/web/src/platform/executor/types.ts` and the route handlers under
    `frontend/web/src/app/**/api/executor/**` (endpoint list mirrored from the
    header of `frontend/web/src/features/executor/client.ts`). It is **the
    compatibility contract for `backend/api` + `backend/workers/executor` +
    `shared/sdk/typescript`**: the Go services being built MUST preserve these
    methods, paths, and response envelopes, and `shared/sdk/typescript` generates its
    types from this file.

    House rules encoded here (PRD §108, §109, house style):

    * **Unknown fields are forbidden to be added by consumers.** Request bodies
      are closed (`additionalProperties: false`); do not extend this contract
      ad hoc — extend this document and regenerate the SDK.
    * **Honest nulls**: fields that may be unknown are `nullable`, never
      faked to `0`.
    * **Credentials are write-only**: secrets are accepted once at connect
      time, sealed server-side, and no route ever returns them. Only masked
      keys (`abc...xyz`) leave the server.
    * Error bodies keep the server's own text (`error`, `detail`, `errors`) —
      clients display `err.message` verbatim.

...
  license:
...
servers:
...
paths:
  /api/executor/preview:
    post:
      tags: [executions]
      operationId: preview
      summary: Dry run — nothing is created, nothing is sent (PRD §80, §98)
      description: |
        Plans the execution against live venue data (instrument, market, fees,
        balances) and returns the preview with conflicts and warnings. Plan
        conflicts are reported INSIDE the 200 preview (`preview.conflicts`);
        they are not an error at preview time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          description: The preview, plus the server-side live-trading kill switch state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewResponse'
        '400':
          description: "Invalid JSON body, request shape, or unresolvable sizing (`error: 'validation'` carries field-named `errors`)."
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/ErrorBasic'
                  - $ref: '#/components/schemas/ErrorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Account unusable (`credential not healthy`) or blocked by position policy (`opposing position exists`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/executor/executions:
    post:
      tags: [executions]
      operationId: createExecution
      summary: Create an execution with an immutable input snapshot (PRD §99)
      description: |
        Validates and plans the request, then persists the execution and its
        immutable creation-time plan. A conflicting plan is **never created**:
        the route answers `409 { error: 'conflict', conflicts, preview }` so the
        caller can show the planner's blocking reasons verbatim. Portfolio
        gates (§73 open risk, §74 daily loss) refuse with `409` before any row
        exists. `mode: 'preview'` is refused — use `POST /api/executor/preview`.
        LIVE mode requires the server-side kill switch (`403` otherwise).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          description: The created record and its immutable creation-time plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateExecutionResponse'
        '400':
          description: "Invalid JSON body / request shape (`error: 'validation'`), or `invalid mode`."
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/ErrorBasic'
                  - $ref: '#/components/schemas/ErrorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Live trading disabled (kill switch is off).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            Plan conflict (the planner's blocking reasons + the preview),
            portfolio-gate refusal (`daily loss guard reached`,
            `max open risk exceeded` — with the numbers that produced the
            refusal as extra fields), unhealthy credential, or an
            existing-position policy refusal.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ConflictError'
                  - $ref: '#/components/schemas/ErrorDetail'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
    get:
      tags: [executions]
      operationId: listExecutions
      summary: Execution history for the session user (PRD §22, §97)
      parameters:
        - name: status
          in: query
          required: false
          description: Optional lifecycle filter; an unknown value is a 400, never silently ignored.
          schema:
            $ref: '#/components/schemas/ExecutionStatus'
      responses:
        '200':
          description: Executions, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListExecutionsResponse'
        '400':
          description: "Unknown `status` value (`{ error: 'status', detail }`)."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/executor/executions/{id}:
    get:
      tags: [executions]
      operationId: getExecution
      summary: One execution and the plan it was created from
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
      responses:
        '200':
          description: The record plus its plan (null only when no plan was persisted — never faked).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/executor/accounts:
    get:
      tags: [accounts]
      operationId: listAccounts
      summary: Connected accounts — masked keys only (PRD §87, §109)
      responses:
        '200':
          description: Credential records (secrets are never returned).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAccountsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [accounts]
      operationId: connectAccount
      summary: BYOK connect — secrets are write-only, sealed server-side
      description: |
        `apiKey`/`apiSecret`/`passphrase` are accepted once here and never
        returned by any route. The live venue probe runs at connect time; a key
        that grants withdrawal permission is refused (400).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectAccountInput'
      responses:
        '200':
          description: The sealed record (masked key) and the venue's probe metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Invalid body, or `withdrawal permission not supported` / validation errors.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/ErrorBasic'
                  - $ref: '#/components/schemas/ErrorDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
  /api/executor/accounts/{id}:
    get:
      tags: [accounts]
      operationId: getAccount
      summary: One connected account — masked key only
      description: |
        Present as a route handler today with no `client.ts` function yet; kept
        in the contract so the Go migration preserves it.
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: The sealed record (masked key) and the venue's probe metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [accounts]
      operationId: deleteAccount
      summary: Revoke a credential (secrets destroyed; record kept masked)
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/executor/settings:
    get:
      tags: [settings]
      operationId: getSettings
      summary: The user's risk profile (PRD §88)
      responses:
        '200':
          description: The effective profile (defaults filled in server-side).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [settings]
      operationId: putSettings
      summary: Partial profile update — unlisted keys fall back to defaults
      description: |
        Strict validation: values are checked, never clamped (a negative
        `maxLeverage` is a 400 with field-named `errors`, not a silently
        corrected write).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskProfilePatch'
      responses:
        '200':
          description: The full stored profile after the merge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: "Invalid JSON body or invalid values (`error: 'validation'`, field-named `errors`)."
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/ErrorBasic'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/executor/emergency:
    post:
      tags: [emergency]
      operationId: emergencyStop
      summary: Emergency stop — stop managed strategies, cancel managed orders
      description: |
        PRD §75: stops all managed strategies and cancels managed open orders
        for the caller (optionally one account). Positions are NEVER closed
        here. Present as a route handler today with no `client.ts` function
        yet; kept in the contract so the Go migration preserves it. An empty
        body means "everything I own".
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmergencyStopInput'
      responses:
        '200':
          description: How many strategies were stopped and orders cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyStopResponse'
        '400':
          description: "`accountId` present but not a string (`error: 'validation'`)."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  # ==========================================================================
  # Treasury read model + CRUD (frontend/web/src/app/(frontend)/api/**)
  # ==========================================================================
...
components:
  parameters:
    ExecutionId:
      name: id
      in: path
      required: true
      description: Execution id.
      schema:
        type: string
    AccountId:
      name: id
      in: path
      required: true
      description: Credential record id.
      schema:
        type: string
  responses:
    Unauthorized:
      description: No session, or the session lacks the `team` tier (fail-closed).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    NotFound:
      description: Not found — or not yours (ownership is never leaked, PRD §108).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBasic'
    UpstreamFailure:
      description: Exchange/venue probe failed; `category` is the retry taxonomy (PRD §78).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...
  schemas:
...
    MarketType:
      type: string
      description: 'Normalized market type. MVP = spot + USDT-margined linear perpetual only.'
      enum:
        - spot
        - linear_perp
    Side:
      type: string
      description: 'Order side. Spot sell-open is refused outside explicit margin support (PRD §89).'
      enum:
        - buy
        - sell
    Intent:
      type: string
      description: What the user wants the order to do to the position.
      enum:
        - open
        - close
        - reduce
...
    MarginMode:
      type: string
      description: 'Margin mode (PRD §92). Product default recommendation is ISOLATED.'
      enum:
        - isolated
        - cross
...
    ExecutionStatus:
      type: string
      description: 'Execution lifecycle (PRD §57).'
      enum:
        - DRAFT
        - CALCULATED
        - VALIDATED
        - READY
        - RUNNING
        - PARTIALLY_FILLED
        - FILLED
        - PAUSED
        - CANCEL_REQUESTED
        - CANCELLED
        - FAILED
        - RISK_STOPPED
        - EXPIRED
        - RECONCILING
        - STOPPED
...
    RiskBreachPolicy:
      type: string
      description: 'When projected risk exceeds the budget (PRD §37). Default RESIZE_THEN_STOP.'
      enum:
        - resize_then_stop
        - pause
        - stop
...
    EntryDefinition:
      description: 'Entry definition (PRD §51).'
      oneOf:
        - type: object
          additionalProperties: false
          required: [type]
          properties:
            type:
              type: string
              enum: [market]
        - type: object
          additionalProperties: false
          required: [type, price]
          properties:
            type:
              type: string
              enum: [limit]
            price:
              type: number
            postOnly:
              type: boolean
      discriminator:
        propertyName: type
    PriceDefinition:
      type: object
      additionalProperties: false
      required: [price]
      properties:
        price:
          type: number
    TakeProfitDefinition:
      type: object
      additionalProperties: false
      required: [price]
      properties:
        price:
          type: number
        fraction:
          type: number
          description: 'Fraction of the position to close at this level (0..1). Absent = full close.'
...
    ExecutionConstraints:
      type: object
      additionalProperties: false
      description: Per-execution hard constraints; absent key = unconstrained.
      properties:
        maxSlippageBps:
          type: number
        maxSpreadBps:
          type: number
        maxPrice:
          type: number
        minPrice:
          type: number
        maxDurationMs:
          type: number
        makerOnly:
          type: boolean
        allowMarketFallback:
          type: boolean
        cancelIfRiskExceeded:
          type: boolean
        stopIfDisconnected:
          type: boolean
    ExecutionRequest:
      type: object
      additionalProperties: false
      required: [accountId, symbol, marketType, side, intent, entry, sizing, execution]
      description: 'The normalized client intent (PRD §51). Unknown fields are rejected — the contract is closed.'
      properties:
        accountId:
          type: string
        symbol:
          type: string
          description: 'Normalized symbol, e.g. `BTC/USDT`.'
        marketType:
          $ref: '#/components/schemas/MarketType'
        side:
          $ref: '#/components/schemas/Side'
        intent:
          $ref: '#/components/schemas/Intent'
        entry:
          $ref: '#/components/schemas/EntryDefinition'
        stopLoss:
          $ref: '#/components/schemas/PriceDefinition'
        takeProfits:
          type: array
          items:
            $ref: '#/components/schemas/TakeProfitDefinition'
        sizing:
          $ref: '#/components/schemas/SizingDefinition'
        leverage:
          $ref: '#/components/schemas/LeverageDefinition'
        marginMode:
          $ref: '#/components/schemas/MarginMode'
        execution:
          $ref: '#/components/schemas/ExecutionDefinition'
        constraints:
          $ref: '#/components/schemas/ExecutionConstraints'
        riskPolicy:
          $ref: '#/components/schemas/RiskBreachPolicy'
        existingPositionPolicy:
          $ref: '#/components/schemas/ExistingPositionPolicy'
        mode:
          $ref: '#/components/schemas/ExecutionMode'
        targetProfit:
          type: number
          description: 'PRD §14 combined constraint: desired profit bound (quote currency), validated TOGETHER with the risk bound.'
        maxRisk:
          type: number
          description: 'PRD §14 combined constraint: maximum-loss bound (quote currency).'
    # ------------------------------------------------------------------
    # Market context captured at planning time
    # ------------------------------------------------------------------
...
    ConnectAccountInput:
      type: object
      additionalProperties: false
      required: [exchange, label, apiKey, apiSecret]
      description: 'BYOK connect body. Secrets are write-only: no route ever reads them back.'
      properties:
        exchange:
          type: string
        label:
          type: string
        apiKey:
          type: string
          writeOnly: true
        apiSecret:
          type: string
          writeOnly: true
        passphrase:
          type: string
          writeOnly: true
    RiskProfilePatch:
      type: object
      additionalProperties: false
      description: 'Partial profile update; unlisted keys fall back to defaults. Values are validated, never clamped.'
      properties:
        defaultRiskMode:
          type: string
          enum: [risk_usd, risk_percent]
        defaultRisk:
          type: number
        maxRiskPerTradePct:
          type: number
        maxOpenRiskPct:
          type: number
        maxDailyLossPct:
          type: number
        maxLeverage:
          type: number
        defaultMarginMode:
          $ref: '#/components/schemas/MarginMode'
        defaultExecutionUrgency:
          $ref: '#/components/schemas/ExecutionUrgency'
    EmergencyStopInput:
      type: object
      additionalProperties: false
      description: 'Optional scope. Omitted/empty body = everything the caller owns (PRD §75).'
      properties:
        accountId:
          type: string
    # ------------------------------------------------------------------
    # Response envelopes — mirror the route handlers / client.ts exactly.
    # ------------------------------------------------------------------
    PreviewResponse:
      type: object
      additionalProperties: false
      required: [preview, liveEnabled]
      properties:
        preview:
          $ref: '#/components/schemas/PreviewResult'
        liveEnabled:
          type: boolean
          description: 'Server-side kill switch (`FUDCOURT_EXECUTOR_LIVE === ''1''`). Never client-side.'
    CreateExecutionResponse:
      type: object
      additionalProperties: false
      required: [execution, plan]
      properties:
        execution:
          $ref: '#/components/schemas/ExecutionRecord'
        plan:
          $ref: '#/components/schemas/ExecutionPlan'
...
    ExecutionDetailResponse:
      type: object
      additionalProperties: false
      required: [execution, plan]
      properties:
        execution:
          $ref: '#/components/schemas/ExecutionRecord'
        plan:
          allOf:
            - $ref: '#/components/schemas/ExecutionPlan'
          nullable: true
          description: 'Null only for a record with no persisted plan; never faked.'
    ListExecutionsResponse:
      type: object
      additionalProperties: false
      required: [executions]
      properties:
        executions:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionRecord'
...
    ListAccountsResponse:
      type: object
      additionalProperties: false
      required: [accounts]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/CredentialRecord'
    AccountResponse:
      type: object
      additionalProperties: false
      required: [account, metadata]
      description: Connect, test and get all answer the sealed record and the venue's probe result.
      properties:
        account:
          $ref: '#/components/schemas/CredentialRecord'
        metadata:
          $ref: '#/components/schemas/AccountMetadata'
    DeleteAccountResponse:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    SettingsResponse:
      type: object
      additionalProperties: false
      required: [profile]
      properties:
        profile:
          $ref: '#/components/schemas/RiskProfile'
    EmergencyStopResponse:
      type: object
      additionalProperties: false
      required: [stopped, cancelledOrders]
      properties:
        stopped:
          type: number
          description: Managed strategies stopped.
        cancelledOrders:
          type: number
          description: Managed open orders cancelled. Positions are never touched.
    # ------------------------------------------------------------------
    # Error shapes — the routes answer with these verbatim bodies.
    # The normalized cross-service error model (code/message/request_id) is
    # shared/contracts/schemas/error-envelope.json.
    # ------------------------------------------------------------------
    ErrorBasic:
      type: object
      additionalProperties: true
      required: [error]
      description: 'Minimal route error — e.g. `{"error": "invalid JSON body"}` or `{"error": "account not found"}`.'
      properties:
        error:
          type: string
          description: The server's own error text; display verbatim.
    ErrorDetail:
      type: object
      additionalProperties: true
      required: [error, detail]
      description: 'Error with a specific reason — e.g. `{"error": "invalid transition", "detail": "..."}` or `{"error": "credential check failed", "detail": "...", "category": "..."}`. Extra numeric fields may carry the numbers behind a refusal (PRD §73/§74).'
      properties:
        error:
          type: string
        detail:
          type: string
        category:
          $ref: '#/components/schemas/ErrorCategory'
    ValidationError:
      type: object
      additionalProperties: true
      required: [error, errors]
      description: 'Field-named validation errors — `{"error": "validation", "errors": ["symbol: ...", ...]}`.'
      properties:
        error:
          type: string
          enum: [validation]
        errors:
          type: array
          items:
            type: string
    ConflictError:
      type: object
      additionalProperties: true
      required: [error, conflicts, preview]
      description: 'A conflicting plan is refused, never created (§117) — `{"error": "conflict", "conflicts": [...], "preview": {...}}`. Conflicts name real fields; show them verbatim.'
      properties:
        error:
          type: string
          enum: [conflict]
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/ConstraintConflict'
        preview:
          $ref: '#/components/schemas/PreviewResult'
    # ------------------------------------------------------------------
    # App-surface error shapes (real route bodies, not the normalized model;
    # that lives in shared/contracts/schemas/error-envelope.json).
    # ------------------------------------------------------------------
...